# docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.errorDto:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
  handler.periodDetailDTO:
    properties:
      duration:
        type: integer
      id:
        type: integer
      max_day:
        type: integer
      type:
        type: string
    type: object
  handler.predictionDTO:
    properties:
      drought_periods:
        type: integer
      normalPeriods:
        type: integer
      optimal_weather_periods:
        type: integer
      periods_detail:
        items:
          $ref: '#/definitions/handler.periodDetailDTO'
        type: array
      rainy_periods:
        type: integer
    type: object
  handler.weatherItemDTO:
    properties:
      day:
        type: integer
      is_peak_day:
        type: boolean
      max_day:
        type: integer
      weather_status:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API to consult weather predictions
  title: "Weather predictor API \U0001F326️"
  version: "1.0"
paths:
  /weather/prediction:
    get:
      parameters:
      - description: Determinate whether return details of periods
        in: query
        name: detailed
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.predictionDTO'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorDto'
      summary: Return summary for predictions
    post:
      responses:
        "202":
          description: ""
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorDto'
      summary: Calculate predictions and store them in database
  /weather/prediction/day/{day}:
    get:
      parameters:
      - description: Day for prediction
        in: path
        name: day
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.weatherItemDTO'
        "400":
          description: For non-positive numbers
          schema:
            $ref: '#/definitions/handler.errorDto'
        "404":
          description: When day was not predict
          schema:
            $ref: '#/definitions/handler.errorDto'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorDto'
      summary: Return prediction for a day
swagger: "2.0"